+++ wiki/Changelog.md
@@ -12,9 +12,11 @@ Font rendering
 
 Window handling
   * Optionally transform Exit to characters, to exit on application-level (#1063).
+  * Options dialog: configurable font and size (~#1059).
 
 Configuration
   * New option ExitCommands (#1063).
+  * New options OptionsFont and OptionsFontHeight (~#1059).
 
 ### 3.4.3 (11 November 2020) ###